Estrae lo spawn procarea e completa sentiero solitario e scudi premio.
Sposta generazione stanze/mob in procarea_spawn.cpp, registra i prototipi
scudo (65100+) al boot e collega touch fontana / enter vortice alla fontana.
Nei messaggi al giocatore il comando resta enter vortice ma viene mostrato
come «entra nel vortice».
